Refactor Python code into a module
This will make testing of the code much easier.
