+++ groovylint/__init__.py
@@ -1,0 +1,1 @@
+"""Main module for groovylint."""
