+++ groovylint/parse_xml.py
@@ -1,0 +1,58 @@
+"""Parse XML reports generated by CodeNarc and print out compiler-like results."""
+
+import xmltodict
+
+
+def _print_violations(package_file_path, violations):
+    for violation in violations:
+        violation_message = f'{violation["@ruleName"]}: {violation["Message"]}'
+        print(f'{package_file_path}:{violation["@lineNumber"]}: {violation_message}')
+
+
+def _print_violations_in_file(package_path, files):
+    for package_file in files:
+        _print_violations(
+            f'{package_path}/{package_file["@name"]}',
+            _safe_list_wrapper(package_file["Violation"]),
+        )
+
+
+def _print_violations_in_package(packages):
+    for package in [p for p in packages if p['@filesWithViolations'] > 0]:
+        # CodeNarc uses the empty string for the top-level package, which we translate to
+        # '.', which prevents the violation files from appearing as belonging to '/'.
+        package_path = package["@path"]
+        if not package_path:
+            package_path = '.'
+
+        _print_violations_in_file(package_path, _safe_list_wrapper(package["File"]))
+
+
+def _safe_list_wrapper(element):
+    """Wrap an XML element in a list if necessary.
+
+    This function is used to safely handle data from xmltodict. If an XML element has
+    multiple children, they will be returned in a list. However, a single child is
+    returned as a dict. By wrapping single elements in a list, we can use the same code to
+    handle both cases.
+    """
+    return element if isinstance(element, list) else [element]
+
+
+def parse(xml_text):
+    """Parse an XML report from CodeNarc."""
+    xml_doc = xmltodict.parse(xml_text)
+
+    package_summary = xml_doc['CodeNarc']['PackageSummary']
+    total_files_scanned = package_summary['@totalFiles']
+    total_violations = package_summary['@filesWithViolations']
+
+    print(f'Scanned {total_files_scanned} files')
+    if total_violations == '0':
+        print('No violations found')
+        return 0
+
+    print(f'Found {total_violations} violation(s):')
+    _print_violations_in_package(_safe_list_wrapper(xml_doc["CodeNarc"]["Package"]))
+
+    return 1
